requirements.txt
boto3==1.40.25
botocore==1.40.25
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
urllib3==1.26.20
pyyaml>=6.0
pandas>=1.5.0
google-generativeai>=0.3.0
